backend/requirements.txt

Django>=4.2
psycopg2-binary
redis
celery
djangorestframework
djangorestframework-simplejwt
gunicorn
